incus/storage_volume: Put big subcommands into their own files
